# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Trentino Stage API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/cv": {
      "post": {
        "summary": "Upload a CV file for the authenticated user",
        "description": "Uploads a single CV file (PDF, DOC, etc.) for the authenticated user. Only one file per user is allowed.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "description": "Session token for authentication"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "The CV file to upload"
                  }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  },
                  "file": {
                    "type": "string",
                    "description": "Base64-encoded file content (not recommended)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File uploaded successfully"
          },
          "400": {
            "description": "No file uploaded or invalid file type"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "User already uploaded a file"
          },
          "405": {
            "description": "Unsupported content type"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "summary": "Download a user's CV file",
        "description": "Returns the CV file for the specified profile if the requester is the owner or a company.",
        "parameters": [
          {
            "in": "query",
            "name": "token",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Session token for authentication"
          },
          {
            "in": "query",
            "name": "profileId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Profile ID of the user whose CV is requested"
          }
        ],
        "responses": {
          "200": {
            "description": "File download stream",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized or invalid session token"
          },
          "404": {
            "description": "No files found"
          },
          "500": {
            "description": "Database connection error"
          }
        }
      },
      "delete": {
        "summary": "Delete the authenticated user's CV file",
        "description": "Deletes the CV file associated with the authenticated user's profile.",
        "parameters": [
          {
            "in": "query",
            "name": "token",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Session token for authentication"
          }
        ],
        "responses": {
          "200": {
            "description": "File deleted successfully"
          },
          "401": {
            "description": "Missing required fields or invalid session token"
          },
          "404": {
            "description": "No files found for this profile"
          },
          "500": {
            "description": "Failed to delete file or database connection error"
          }
        }
      },
      "put": {
        "summary": "Replace the authenticated user's CV file",
        "description": "Replaces the existing CV file for the authenticated user with a new one. Only one file per user is allowed.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "description": "Session token for authentication"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "The new CV file to upload"
                  }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  },
                  "file": {
                    "type": "string",
                    "description": "Base64-encoded file content (not recommended)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File uploaded successfully"
          },
          "400": {
            "description": "No file uploaded or invalid file type"
          },
          "401": {
            "description": "Missing required fields or invalid session token"
          },
          "405": {
            "description": "Unsupported content type"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}